Organise Various_scripts for incorporation
The files included here are copy pasted directly from my Various_scripts repository. All of their various functions are wanted to be implemented into variantopia. With this in mind, I will need to consider the underlying data structures / classes that will enable for all these functions to be handled with some level of flexibility for future additions.